+++ src/main/java/qna/repository/AnswerRepository.java
@@ -1,11 +1,12 @@
-package qna.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package qna.repository;
 
 import java.util.List;
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.Answer;
 
 public interface AnswerRepository extends JpaRepository<Answer, Long> {
+
     List<Answer> findByQuestionIdAndDeletedFalse(Long questionId);
 
     Optional<Answer> findByIdAndDeletedFalse(Long id);
