+++ src/main/java/qna/repository/DeleteHistoryRepository.java
@@ -1,6 +1,8 @@
-package qna.domain;
+package qna.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.DeleteHistory;
 
 public interface DeleteHistoryRepository extends JpaRepository<DeleteHistory, Long> {
+
 }
