+++ src/main/java/qna/repository/UserRepository.java
@@ -1,9 +1,10 @@
-package qna.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package qna.repository;
 
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
+
     Optional<User> findByUserId(String userId);
 }
